renamed ProjectsFilter to index for clarity
